+++ src/variable.zig
@@ -10,9 +10,10 @@ pub const pawn_threat_rook = 0;
 pub const pawn_threat_queen = 0;
 
 pub const pawn_defend_king = 5;
-pub const pawn_isolated = 50;
-pub const pawn_doubled = 20;
+pub const pawn_isolated = 30;
+pub const pawn_doubled = 15;
 pub const pawn_blocked = 10;
+pub const pawn_protection = 30;
 
 pub const bishop_pair = 10;